Use the official CMSSW repository in AFS as we moved it from vocms146. Also have an option for each repository.
This is the first patch for our patches-mailing list at cms-cond-dev@cern.ch and it is intended as a real example.

It changes deploy.py to use the official "CMSSW" repository (i.e. the one that keeps the /data/cmssw folder) that we previously had on vocms146 temporarily. Also includes an option for each repository (cmssw was "special" before, now all the three repositories stay in the same folder).

Feel free to comment! I am wearing the developer hat here, tomorrow I will wear the maintainer hat (pusher hat) and apply the patch to the repo.

Signed-off-by: Miguel Ojeda <mojedasa@cern.ch>