Use virtualHost's IP instead of the hostname's.
This allows to generate the same files in different machines,
which is better for consistency, for scp'ing the files
to the frontends from other machines. Also useful if in
the future we have several network interfaces, etc.

Originally the current hostname was used so that the .conf
files could be directly tested in a fake frontend
(e.g. private VM).

Signed-off-by: Miguel Ojeda <mojedasa@cern.ch>