Use an Apache frontend for private development, like in dev/int/pro.
This reflects more the dev/int/pro environment and,
more importantly, allows to query other services using
absolute paths in both private and dev/int/pro instances.
For instance, a /libs/jquery.js link in a popcon's page
would not work in a private instance as it would wrongly point
to https://myvm.cern.ch:8082/libs/, i.e. popcon's CherryPy would
reply to the request, failing.

This commit also updates the links lists in docs' front page
and in the keeper status command.

In addition, these changes require to open the 80 (HTTP)
and 443 (HTTPS) ports in the private machines in the deploy
script. Opening the full range of ports is no longer required,
so it has been removed.

Moreover, the deploy script now calls makeApacheConfiguration.py
to set up the private Apache frontend (which in turn is generated
like an exact copy of the -dev frontend's configuration with
a different hostname and no Shibboleth). Also, sets
SELinux's httpd_can_network_connect to 'on', which allows
Apache run in enforcing mode (default for CERN VMs).
Finally does a graceful restart of Apache.

Signed-off-by: Miguel Ojeda <mojedasa@cern.ch>