New common/api.py: CherryPy class decorator for generating the api() and apih() methods.
Use it in the gtList, payloadInspector and recordsProvider services
and remove the old api() method if it existed. CMSCONDDB-8

In addition, add the service.py's getPrettifiedJSON() function to
(only) prettify JSON and make setResponseJSON() use it. This is needed
for pregenerating the JSON in api.py to avoid having to prettify it
every time api() is called.

Kostas provided the original patch.

Signed-off-by: Kostas Tamosiunas <kostast@cern.ch>
Signed-off-by: Miguel Ojeda <mojedasa@cern.ch>