Always strip the domain.
In some machines the hostname is the FQDN (e.g. vocms*),
in others is not (e.g. private VMs).

Previously it was done only in getCurrentFrontend(), therefore
rename getCurrentFrontend() to getHostname() and use it everywhere.

This prevents double .cern.ch.cern.ch strings when deploying
privately on vocms* machines (i.e. using '-f private' for debugging).

Signed-off-by: Miguel Ojeda <mojedasa@cern.ch>