status() now uses the new formatTable() to format its output.
formatTable() automatically calculates the maximum width
of each column and fills as appropiate. Also allows to specify
custom column and row separators and the filler character
for spacing.

This will allow to include other information/columns (e.g. jobs)
in the table easily.

Also now a header is printed to identify the columns.

Signed-off-by: Miguel Ojeda <mojedasa@cern.ch>