Add common http.py to query HTTP(S) URLs and use it in tests.
This mainly adds a HTTP class for general querying in http.py,
with a custom HTTPError class, simplifying and abstracting
our HTTP requests (now we use pycurl instead of urllib{,2}).

In turn, the HTTP class is used by the also new HTTPService class
in service.py which mimics the old query*() functions, i.e. it is
used to query your own service.

In turn, the HTTPService class is used in the service.py's
specialized TestCase for our services.

Signed-off-by: Miguel Ojeda <mojedasa@cern.ch>