Fixed dropBoxBE's logging.
Now the stream handler (i.e. logging to the console, and in turn
to the web logs) is handled by 'import service' as usually done
in other services. This means a copy of all logging goes to the web logs.

Then, TeeFile just needs to care about logging to a file.

The name of the logger was removed from the logs, since now
the logs are properly divided.

Also, TeeFile now removes all handlers properly, so we do not get
repeated entries, which may occur since FileDownloader creates
a TeeFile several times with the same name; and also now the Dropbox
instance is kept forever alive, so we need to reset them to avoid
an ever-increasing number of handlers.

Signed-off-by: Miguel Ojeda <mojedasa@cern.ch>