Move shutdown()'s logic to processAllFiles() and recall errors for DONE_WITH_ERRORS.
Since we now run forever with the same Dropbox instance, we need to
report the status to the server, upload the logs and back them up
in processAllFiles(). Therefore, remove shutdown() and put its logic
into processAllFiles(). Some of its logic was not needed in any case
anymore (like deleting loggers).

Signed-off-by: Miguel Ojeda <mojedasa@cern.ch>