Wait for the delay in server.py where we do the main loop.
This simplifies the logic for test.py. Since:

  * Processing the files should be quick.
  * We do not have a lot of requests.
  * We will run in continuous mode.

it means it should not hurt to wait when we have done a bunch,
even if in-between new files arrive.

We can change this later if we need to process files as fast
as possible; while we rethink how to update the runLog.
This depends on when we (re)create StatusUpdater.

Signed-off-by: Miguel Ojeda <mojedasa@cern.ch>