In tests, use INFO logging level by default.
Even when running in a private instance, usually the tests
do not contain really complex logic to debug, so we just
want them to pass (i.e. most of the time we are adding tests,
not debugging the testing logic), so by default we run in INFO
level. Otherwise, there is a lot of output from database queries
and so on for some services.

By the way, we would need a second debug level for really low-level
debugging, like printing the database queries and so on.

Signed-off-by: Miguel Ojeda <mojedasa@cern.ch>