For some reason condDB.RDBMS(authpath) gets 'overwritten' by dropBoxBE.reprocess().
If we instantiate one time conditionDatabase.condDB.RDBMS(conf.authpath)
and then use it after dropBoxBE.reprocess() runs, the authentication is
lost and the connection to the dropbox database fails. It looks like
there is something 'global' with side-effects in Dropbox.py or
conditionDatabase.py. Maybe the FWIncantation() has global state?

This conclusion was reached since running/testing the code skipping
the uploading and processing of the files does not trigger the problem,
i.e. if dropBoxBE.reprocess() is commented out, the rdbms works fine.

However, when processing the files, it does not. The problem is easily
reproducible running dropBox runs 420..440, i.e. this change:

 diff --git a/dropBoxBE/replay.py b/dropBoxBE/replay.py
 index 4f77ea9..aaf33e5 100755
 --- a/dropBoxBE/replay.py
 +++ b/dropBoxBE/replay.py
 @@ -164,7 +164,7 @@ def main():
      _fwLoad = conditionDatabase.condDB.FWIncantation()

      i = 0
 -    for runTimestamp in sortedDropBoxRuns:
 +    for runTimestamp in sortedDropBoxRuns[420:440]:
          i += 1
          logging.info('[%s/%s] %s: Replaying run...', i, len(dropBoxRuns), runTimestamp)

Therefore and for the moment, to be able to run the replay,
I instantiate every time the RDBMS.

Signed-off-by: Miguel Ojeda <mojedasa@cern.ch>