When porting metadata, change since == 1 to None for some files that really meant it.
In the old dropBox we allowed either None or a since which was smaller
that the one in the data, which made the one in the data being used.
Since in the new dropBox we only allow None to mean 'take the one from
the data', we overwrite it here so that we do not discard the file.

Signed-off-by: Miguel Ojeda <mojedasa@cern.ch>