PyCurler now uses http.py internally, which adds retries and timeout support.
This tries to make the backend more reliable to network problems
and, until it is solved, the Apache-CherryPy 502 Bad Gateway problems.

Also, this changes the URLs to the frontend to point directly
to Apache (i.e. the frontends), since from the online network
we cannot access ports through the proxy which are not 80/443
(i.e. 8095 for dropBox) and there is not a DNS entry for cmsproxy
in the GPN side.

Signed-off-by: Miguel Ojeda <mojedasa@cern.ch>