Cache the upload script in memory and do not serve it as a file.
Since we will put the upload script in the TWiki, we can serve it as a
string and cache it in memory, instead of serving it as a file every
time. Another solution is to use a file object (which must be recreated
every time since CherryPy closes it after using it).

Signed-off-by: Miguel Ojeda <mojedasa@cern.ch>