Fix empty globalLogs: The handlers were being removed by the new object being created.
This was a problem with self.logger = TeeFile(...). The handlers were
cleaned up on __del__, but since the new object TeeFile is created
before is assigned to self.logger (which removes self.logger *after*),
and getLogger() returns the same logger, the handlers were being removed
after the new object added the new ones, which explains why it worked
the first time but not the consecutive ones.

The downloaderLog was working since it was deleted before and the
__del__ was called, although we should not rely on that: therefore
we are calling removeHandlers() for all loggers (downloadLog, globalLog
and the one in FileDownloader) explicitly now before logs are sent,
which also solves the original problem.

If the processing is not successful is still OK since removeHandlers()
in __init__ will take care of them (so no ever-increasing logs).
Basically, this is like a close() for a normal file and we need to
explicitly do it.

In any case I would suggest creating a class supporting the with
statement and wrapping all the processing in it.

This commit also includes some debugging lines in case of future problems.

Signed-off-by: Miguel Ojeda <mojedasa@cern.ch>