Use file-based sessions in dropBox to avoid problems in dropBoxBE.
This change makes sure dropBoxBE does not break when it loses
the session.

While we should make dropBoxBE retry in case of lost session
(i.e. re-signIn), this would work as a safety net in case
the retry code has a bug, so it is good to keep the sessions in files.

In any case, we should not lose the files in /data/files (or else we
would have bigger problems than lost sessions if services start storing
them more critical data, e.g. the dropBoxBE can use it for the work area
i.e. NewOfflineDropBoxBaseDir).

Signed-off-by: Miguel Ojeda <mojedasa@cern.ch>