Fix invalid connection string check and separate it from the unsupported destination database one.
If the connection string was invalid we were returning None from
getProtocolServiceAndAccountFromConnectionString() so we could not get
the 'service' key; fixed now.

Also, the invalid connection string check can be done independently
of the unsupported destination database one, which allows to test the
first in the private instances.

Both of these were catched by making the tests pass again (which goes in
the next commit).

Signed-off-by: Miguel Ojeda <mojedasa@cern.ch>