Trap SIGTERM to wait until we have finished processing all files.
This allows to easily stop the emails service with the keeper
or kill without worrying (too much) about breaking a request.

Still it would be probably good to implement another signal-free
to stop the process, like checking for a file.

I will make that improvement while adding support in the keeper
for 'backend' services.

Signed-off-by: Miguel Ojeda <mojedasa@cern.ch>