Redirect /monitoring/ to /monitoring/check_mk/ to prevent issues with ../ URLs.
The other way was to enable also a location for /pnp4nagios and any
others, which looked too hacky. I also tried to rename our OMD site
to "monitoring", so that all URLs are the same, but "omd mv" does not
"just work" (tries to find a init.d script that is not created).
I will try to fix that.

The alternative is to recreate the site and copy the configuration.
However, that does not work, since you need to redo all the
configuration, URLs, paths, etc. in many places.

Signed-off-by: Miguel Ojeda <mojedasa@cern.ch>