Disallow the online user for user functions.
Now checkSignedIn does not match the online user:

  all = signIn() allows cms-cond-dropbox and the online user.
  checkSignedIn = all - online user = cms-cond-dropbox
  checkSignedInOnline = only online user

This means checkSignedIn is now the complement set of
checkSignedInOnline in the set of all users.

This prevents us (the developers) of using the online user account
(popconpro) to upload things with a non-CERN user which breaks
the emailing system. In any case, we are also fixing the emails
in case CERN LDAP breaks (for this or other reasons) in another commit.

Signed-off-by: Miguel Ojeda <mojedasa@cern.ch>