Rewritten manageApacheWorkers.py: Simplified and optimized.
Now a separate class BalancerManager takes care of the interface
to Apache's balancer-manager. A refresh() method parses the main
page to get all available balancers, their workers and their status.

URLs are not built anymore (although we still rely on some structure
in the HTML for the regexps...). This avoids the "http" FIXME
of a previous commit for the gtc service and also removes
the dependency on config.py, since we do not need the ports either.

In addition, now all regexps are called once for all the text.
Overall, the script is simpler and should be easier to update
for future versions of balancer-module.

manageBackend() now just needs to loop throught the gathered
information exposed by BalancerManager and trigger any action
as required.

Signed-off-by: Miguel Ojeda <mojedasa@cern.ch>