Limit results to the last week and add search feature to browse through the rest.
Limiting the results to the last week will improve performance for the
most common case, and, to compensate, we add the search feature to be
able to browse the rest of the table.

The search feature is able to filter by any specified column and is
*not* case sensitive. It also allows to specify a category for the
statusCode (Success, Warning, Failure or Any). The only mandatory fields
is the date range.

In addition, the page looks exactly the same as the normal one when
searching, and all the parameters are carried in the URL so that
it can be copied into other places (e.g. emails) for reference.
The page fills the form with the search values to allow for easy
modification as well.

Also, this moves the DropBox class from the logs/logs.py to the logs/dropBox/logs.py
where it should belong.

Signed-off-by: Miguel Ojeda <mojedasa@cern.ch>