Overwrite error code with warning in case there are exceptions in the long tail.
If there was already an error in the times/gaps (the error coming
from isTimeConsistent) we keep always the error. If not, and there is
no error in the long tail, we keep the 0 (no error). However, if there
was an error in the long tail, we mark it as a warning (code 2).

Signed-off-by: Salvatore Di Guida <salvatore.di.guida@cern.ch>
Signed-off-by: Miguel Ojeda <mojedasa@cern.ch>