Make the OfflineDropBox tab's icon useful.
A red icon means there was an error in the logs *or*
that the DropBox did not run for a long time
(2 hours, since it is the minInterval -- the tolerance
does not kick since the DropBox runs every 10 minutes).

A orange/warning icon means there was a warning in the logs.

Signed-off-by: Salvatore Di Guida <salvatore.di.guida@cern.ch>
Signed-off-by: Miguel Ojeda <mojedasa@cern.ch>