Add a single retry to signIn if we lost the session to the frontend.
We can identify this because the server returns 404 Not Found
in case you are not logged in.

This prevents having to manually finish uploads when something
goes really slow, e.g. Oracle queries, like it happened
on 2013-02-06 when Oracle was heavily loaded by another session
and it took more than 20 minutes to finish several exportations.

The script emulateDropBoxBE.py was written to manually finish
an upload since those problems happened. With this fix we should
not need to use it in this case.

Signed-off-by: Miguel Ojeda <mojedasa@cern.ch>