Check fcsr's availability and validity in the frontend (if required).
If an upload asks for a synchronization to prompt or pcl (in export
or in the dependencies), the frontend now checks whether it can reach it
(i.e. network issues, Tier0's VM down...) and whether it is returning
a valid value (i.e. not None -- which happens when they switch off
Tier0 but the data service is still running).

If this is not true, it returns immediately an error to the user
explaining the situation and who to contact, depending on whether
we are on data taking or not.

Still, we need to do the same checking in the backend, since the Tier0
value may change in-between and Tier0 could switch off their data
completely (because at the moment we always ask for the fcsr
in the backend at the start in all cases, even if not needed, which
would crash the backend even if the request is not related to
prompt or pcl). See next commit for backend's changes.

Signed-off-by: Miguel Ojeda <mojedasa@cern.ch>