Handle unreachable and invalid fcsr in the backend.
Previously the backend crashed if the fcsr was not available
or if it was invalid. This was OK since during data taking
it was supposed to be correct at all times.

Now we try to get the fcsr and store it. If we need to use it
now the file is marked as failed with the respective constant.
The fcsr is used when an upload synchronizes to prompt or pcl,
but the code will behave the same if the value is used in any case.

If this happens, the offline DB shifter will see the problem
and receive the usual SMS. He can act accordingly and timing
should not be an issue since for prompt and pcl we have got
48 hours. This will be used in October/November data taking.

See previous commit for details on checks the frontend side
to prevent even reaching this situation in most cases.

In addition, the DB's constraint was relaxed to allow to save
those values. This is meant to be able to store the HLT run number
regardless of problems with the fcsr.

Signed-off-by: Miguel Ojeda <mojedasa@cern.ch>