Redirects-only services should also redirect the URLs without slashes.
Services with proxies and redirects worked fine since a redirect
rule without slash is added for proxied services; but services
with only redirects did not have the rule for the non-slash case.

This makes /easymon_online and /frontier_online work (i.e. the same
as /easymon_online/ and /frontier_online/ before).

Signed-off-by: Miguel Ojeda <mojedasa@cern.ch>