Pick the DB connection dictionary from the secrets instead of authentication.xml.
The authentication.xml file was not updated from the official one.

While this does not solve the original problem, moves the secrets into
the proper place so that are centralized and, in this way, in the future
we may implement another way to store them (e.g. in the DB within
the key-based authentication system).

Signed-off-by: Miguel Ojeda <mojedasa@cern.ch>