Remove separation of Easymon's boxes.
In the beginning I did not add separation, but I did not have
borders either, so one could not see how big the boxes were (weights).

After that I implemented the borders, and therefore now we can merge
the boxes again (i.e. no separation), which allows to gain a little bit
of resolution in small screens and looks better conceptually, since
the status is the sum of all of the boxes.

Signed-off-by: Miguel Ojeda <mojedasa@cern.ch>