Generate ssl.conf and shib.conf automatically as well.
These two were the last parts that were not automatically generated
by our scripts to configure Apache.

The files themselves have been stripped of comments and placed in
the same place they would have been imported.

Also, since now they are not needed and, in fact, they should not be
imported, they are disabled in private machines like other files
previously. However, now all the disabling is done in
makeApacheConfiguration.py and is done via removing all *.conf files
(backed up) in the folder, instead of a whitelist.

This will improve consistency and ease the port to OS X, as well as
the move of all the configuration to /data/httpd and the 'per-group'
or 'per-project' httpd in the future.

Signed-off-by: Miguel Ojeda <mojedasa@cern.ch>