Remove historical SetEnvIf meant for MSIE <= 5.
Even Internet Explorer 6 has fixed the keepalive issues,
and therefore, we could just use the ssl-unclean-shutdown part.

However, since this was just in the _default_ virtual host,
we were not using it and nobody complained. Therefore,
let's just remove it altogether.

This will also allow us to remove the setenvif module
in the next commit.

Signed-off-by: Miguel Ojeda <mojedasa@cern.ch>