Fix daylight saving time issue.
The dropBox frontend was unable to insert rows in the runLog because
the constraint that checks that the modificationTimestamp is equal or
higher than the creationTimestamp was failing.

The problem is that the application reuses the DB connection/session
(as should) but the default value for the modificationTimestamp column
was CURRENT_TIMESTAMP, which does follow daylight saving time, *but* only
when the session is created, i.e. it does not get automatically updated
unless the session is recreated, which means it still was in winter time
after summertime took place.

This triggered a problem in the comparison because, in the runLog's query
to update the status, the client (the dropBoxBE) specifies the
creationTimestamp, instead of using the default value (which would have
worked since it would have been also in winter time but we would have
had wrong winter times in summer anyway -- and would be confusing after
the next restart of the dropBoxBEs which would move the times forward).

After the runLog query fails, other inserts (e.g. to fileLog) failed
since the parent key was not present in the runLog, which is why we did not
have the files' logs either.

The problem could be solved temporarily (until the next daylight saving
time change) by restarting the session (i.e. restarting the dropBoxes),
but the proper fix is to change CURRENT_TIMESTAMP to SYSTIMESTAMP,
which is the server time instead of the session time and this one
*does* follow daylight saving time as one would expect.

This commit just changes the original dataModel.sql file for future
reference but it is not required to fix the issue. The issue was fixed
running the commands shown below [1] which alter the tables in
the production DB. We did so for all the tables and triggers
(otherwise we would have wrong dates in the tables even if no
constraint fail).

Since we did not restart the dropBoxes before running the commands,
after applying the changes the dropBoxes started working properly
right away, and therefore the change was itself another test.

A final upload to test the full chain was also done.

[1]

alter table files modify (creationTimestamp default SYSTIMESTAMP,
modificationTimestamp default SYSTIMESTAMP);
create or replace trigger tumFiles
before update on files
for each row
begin
        :new.modificationTimestamp := SYSTIMESTAMP;
end;
/

alter table runLog modify (modificationTimestamp default SYSTIMESTAMP);
create or replace trigger tumRunLog
before update on runLog
for each row
begin
        :new.modificationTimestamp := SYSTIMESTAMP;
end;
/

alter table fileLog modify (creationTimestamp default SYSTIMESTAMP,
modificationTimestamp default SYSTIMESTAMP);
create or replace trigger tumFileLog
before update on fileLog
for each row
begin
        :new.modificationTimestamp := SYSTIMESTAMP;
end;
/

alter table fileAcks modify (creationTimestamp default SYSTIMESTAMP,
modificationTimestamp default SYSTIMESTAMP);
create or replace trigger tumFileAcks
before update on fileAcks
for each row
begin
        :new.modificationTimestamp := SYSTIMESTAMP;
end;
/

alter table emails modify (creationTimestamp default SYSTIMESTAMP,
modificationTimestamp default SYSTIMESTAMP);
create or replace trigger tumEmails
before update on emails
for each row
begin
        :new.modificationTimestamp := SYSTIMESTAMP;
end;
/

Signed-off-by: Miguel Ojeda <mojedasa@cern.ch>