Support all parameters in balancer-manager by inspecting the admin form.
The parameters may change with different Apache versions (both names
and values!). This was changed last time to support the new ones
with the change from 2.2.3 to 2.2.15; however, looks like with the
upgrade to SLC6.4 now the original ones are again used.

To prevent future problems if in a future SLC version the new ones are
used, now we support both sets of parameters. Of course, this does not
solve the issue of facing new names/values in newer versions.

Signed-off-by: Miguel Ojeda <mojedasa@cern.ch>