Escape quotes in html.escape().
xml.sax.saxutils.escape() does not escape quotes by default,
only the < > & characters. While this prevents from opening
new tags like <script>, it does not prevent modifying
the attributes of a tag if it is called to escape one of
the attributes of a tag (instead of content inside a tag),
which can be used to modify the page layout or trigger
inlined scripts (e.g. onclick attribute).

See http://wiki.python.org/moin/EscapingHtml

Signed-off-by: Miguel Ojeda <mojedasa@cern.ch>