Revert gtc's SERVICE_ACCOUNT_NAMES setting to point to webcondvm2.
5be9256 removed the last references to webcondvm2, in
SERVICE_ACCOUNT_NAMES and SERVICE_GT_INFO_UPDATE. However, it turns out
that get_connectionName does not work properly in the official
payloadInspector (never did, and there were no users of it).

Since the gtc already uses the secrets to fetch the accounts for other
purposes, there is no need to keep get_connectionName in
payloadInspector. Dimitar will use the secrets as well to fill that
information -- which is more consistent as well and centralizes the
knowledge of the accounts.

Therefore, until this issue is resolved, the previous change is
reverted. This commit is not an exact revert of the other one since
removing SERVICE_GT_INFO_UPDATE is fine nevertheless.

See CMSCONDDB-55 for more details.

Signed-off-by: Miguel Ojeda <mojedasa@cern.ch>