Make popconSQL.py use common/database.py and major cleanup.
Previously every single time a page was loaded a connection
was created to Oracle. Now only one is used, and bind variables
are used. Both improvements should increase performance.

This also includes:

  * Simplified functions' logic.
  * Simplified some queries.
  * Removed unused variables.
  * Removed commented out code.
  * Removed tabs (whitespace).

The previous commits that removed unused code allowed to minimize
the amount of changes required in this one; otherwise a lot of
extra code would have need a cleanup/rewrite as well.

The changes were tested via visual inspection via the browser
plus, for pages with a lot of data, making diffs of the output
returned by curl for the previous and new version.

Also, the output from the popcon.py monitoring script
was tested to match the previous and new version.

Signed-off-by: Miguel Ojeda <mojedasa@cern.ch>