Unlimited width on ps's output.
The rules of ps regarding trimmed output when piped
are not that well-defined, quoting from the manual:

''' If ps can not determine display width, as when output
    is redirected (piped) into a file or another command,
    the output width is undefined. (it may be 80, unlimited,
    determined by the TERM variable, and so on).
    '''

This made common/service.py's _getPIDs() misbehave in some cases,
since the command-line column (last of 'ps aux') was trimmed.

This, in turn, caused isAnotherInstanceRunning() to return
false negatives.

I could not reproduce it in keeper's _getPIDs() which uses
ps -eo. In any case, within the terminal it still happens,
so the -ww option was included as well there.

Signed-off-by: Miguel Ojeda <mojedasa@cern.ch>