Enable KeepAlive with small values (for the moment).
Enabling KeepAlive will help a lot with pages with many
requests like gtList and payloadInspector and hides to
some degree the throttling problem.

In addition it helps a lot since we use SSL everywhere.

The values I chose are more than enough to help all
services improve their performance and, being conservative,
they should not explode Apache's memory usage. If later on
we see the memory usage is still OK we may increase the values.

For some reason it is disabled by default in RHEL,
but for our usage it is way better to enable it to a small degree.

Signed-off-by: Miguel Ojeda <mojedasa@cern.ch>